+++ p-ata/Cargo.toml
@@ -94,9 +94,6 @@ spl-associated-token-account-client = { version = "2.0.0", optional = true }
 bincode = { version = "1.3.3", optional = true }
 solana-sdk-ids = { version = "2.2.1", optional = true }
 test-case = { version = "3.3.1", optional = true }
-# End optional additions
-
-# ---------------------- std-only optional dependencies ----------------------
 solana-account = { version = "2.2.1", optional = true }
 solana-instruction = { version = "2.3.0", optional = true }
 solana-logger = { version = "2.3.0", optional = true }
@@ -104,17 +101,13 @@ solana-pubkey = { version = "2.2.1", features = ["curve25519"], optional = true 
 solana-sysvar = { version = "2.2.2", optional = true }
 solana-system-interface = { version = "1.0.0", git = "https://github.com/rustopian/system.git", branch = "create-prefunded-account", optional = true }
 solana-program-test = { version = "3.0", git = "https://github.com/rustopian/agave.git", branch = "create-prefunded-account", optional = true }
-
 mollusk-svm = { version = "0.4.1", git = "https://github.com/rustopian/mollusk.git", branch = "rustopian/create-prefunded-account", features = ["all-builtins"], optional = true }
 mollusk-svm-bencher = { version = "0.4.1", git = "https://github.com/rustopian/mollusk.git", branch = "rustopian/create-prefunded-account", optional = true }
-
 curve25519-dalek = { version = "4.1.3", default-features = false, optional = true }
-
 spl-token = { version = "8.0.0", features = ["no-entrypoint"], optional = true }
 spl-token-2022 = { version = "8.0.1", features = ["no-entrypoint"], optional = true }
 spl-token-group-interface = { version = "0.6.0", optional = true }
 spl-token-metadata-interface = { version = "0.7.0", optional = true }
-
 comfy-table = { version = "7", optional = true }
 strum = { version = "0.27.1", features = ["derive"], optional = true }
 serde = { version = "1.0", features = ["derive"], optional = true }
@@ -123,6 +116,7 @@ bs58 = { version = "0.4.0", optional = true }
 colored = { version = "2.0", optional = true }
 itertools = { version = "0.12", optional = true }
 rstest = { version = "0.26.1", optional = true }
+spl-token-2022-interface = "0.0.0"
 # ---------------------------------------------------------------------------
 
 [dev-dependencies]
