--- p-ata/src/tests/token_account_len/mod.rs
@@ -1,6 +1,0 @@
-#![cfg(test)]
-
-mod test_account_length_limits;
-mod test_extension_size_exhaustive;
-mod test_extension_size_validation;
-mod test_extension_utils;
