--- p-ata/src/tests/token_account_len/test_extension_utils.rs
@@ -1,369 +1,0 @@
-#![cfg_attr(feature = "std", allow(dead_code, unused_imports))]
-
-#[cfg(any(test, feature = "std"))]
-use {
-    spl_token_2022::{
-        extension::{
-            default_account_state::DefaultAccountState, group_pointer::GroupPointer,
-            interest_bearing_mint::InterestBearingConfig, metadata_pointer::MetadataPointer,
-            mint_close_authority::MintCloseAuthority, non_transferable::NonTransferable,
-            pausable::PausableConfig, permanent_delegate::PermanentDelegate,
-            transfer_fee::TransferFeeConfig, transfer_hook::TransferHook,
-            PodStateWithExtensionsMut,
-        },
-        pod::PodMint,
-    },
-    spl_token_group_interface::state::{TokenGroup, TokenGroupMember},
-    spl_token_metadata_interface::state::TokenMetadata,
-    std::{format, string::String, vec::Vec},
-};
-
-#[derive(Debug, PartialEq)]
-pub enum ExtensionCategory {
-    Include,
-    AccountOnly,
-    Skip,
-}
-
-/// If new variants are added to ExtensionType, this function will fail to compile
-/// until the new variants are explicitly handled. Note this ignores the program's
-/// anticipated "`PlannedZeroAccountDataLengthExtension`"
-pub fn categorize_extension(ext: spl_token_2022::extension::ExtensionType) -> ExtensionCategory {
-    use spl_token_2022::extension::ExtensionType::*;
-    match ext {
-        Uninitialized => ExtensionCategory::Skip,
-
-        // Mint extensions that can be initialized independently
-        TransferFeeConfig
-        | NonTransferable
-        | TransferHook
-        | Pausable
-        | DefaultAccountState
-        | InterestBearingConfig
-        | MetadataPointer
-        | GroupPointer
-        | GroupMemberPointer
-        | MintCloseAuthority
-        | PermanentDelegate
-        | ScaledUiAmount
-        | TokenMetadata
-        | ConfidentialTransferMint
-        | ConfidentialTransferFeeConfig
-        | TokenGroup
-        | TokenGroupMember
-        | ConfidentialMintBurn => ExtensionCategory::Include,
-
-        // Account-only extensions - these shouldn't be in mint data
-        TransferFeeAmount
-        | ConfidentialTransferAccount
-        | ImmutableOwner
-        | NonTransferableAccount
-        | TransferHookAccount
-        | ConfidentialTransferFeeAmount
-        | PausableAccount
-        | MemoTransfer
-        | CpiGuard => ExtensionCategory::AccountOnly,
-    }
-}
-
-/// Create mint data with specific extensions using token-2022's official methods
-pub fn create_mint_data_with_extensions(
-    extension_types: &[spl_token_2022::extension::ExtensionType],
-) -> Vec<u8> {
-    use spl_token_2022::extension::{BaseStateWithExtensionsMut, ExtensionType};
-    use std::string::String;
-    use std::{vec, vec::Vec};
-
-    let required_size = if extension_types
-        .iter()
-        .any(|ext| matches!(ext, ExtensionType::TokenMetadata))
-    {
-        // Calculate length for all sized extensions first, then add buffer for TokenMetadata
-        let mut size = ExtensionType::try_calculate_account_len::<spl_token_2022::state::Mint>(
-            &extension_types
-                .iter()
-                .copied()
-                .filter(|e| !matches!(e, ExtensionType::TokenMetadata))
-                .collect::<Vec<_>>(),
-        )
-        .expect("calc len for sized subset");
-
-        const TOKEN_METADATA_VALUE_LEN_ESTIMATE: usize = 500;
-        size += TOKEN_METADATA_VALUE_LEN_ESTIMATE + 4; // value + TLV header
-        size
-    } else {
-        ExtensionType::try_calculate_account_len::<spl_token_2022::state::Mint>(extension_types)
-            .expect("Failed to calculate account length")
-    };
-
-    let mut data = vec![0u8; required_size];
-
-    let mut mint = match PodStateWithExtensionsMut::<PodMint>::unpack_uninitialized(&mut data) {
-        Ok(mint) => mint,
-        Err(e) => panic!(
-            "Failed to unpack mint for extensions {:?}: {:?}",
-            extension_types, e
-        ),
-    };
-
-    mint.base.mint_authority = Default::default();
-    mint.base.supply = 0u64.into();
-    mint.base.decimals = 6;
-    mint.base.is_initialized = true.into();
-    mint.base.freeze_authority = Default::default();
-
-    for extension_type in extension_types {
-        match extension_type {
-            ExtensionType::TransferFeeConfig => {
-                let extension = mint
-                    .init_extension::<TransferFeeConfig>(true)
-                    .expect("Failed to init TransferFeeConfig");
-                extension.transfer_fee_config_authority = Default::default();
-                extension.withdraw_withheld_authority = Default::default();
-                extension.withheld_amount = 0u64.into();
-            }
-            ExtensionType::NonTransferable => {
-                mint.init_extension::<NonTransferable>(true)
-                    .expect("Failed to init NonTransferable");
-            }
-            ExtensionType::TransferHook => {
-                let extension = mint
-                    .init_extension::<TransferHook>(true)
-                    .expect("Failed to init TransferHook");
-                extension.authority = Default::default();
-                extension.program_id = Default::default();
-            }
-            ExtensionType::Pausable => {
-                let extension = mint
-                    .init_extension::<PausableConfig>(true)
-                    .expect("Failed to init PausableConfig");
-                extension.authority = Default::default();
-                extension.paused = false.into();
-            }
-            ExtensionType::DefaultAccountState => {
-                let extension = mint
-                    .init_extension::<DefaultAccountState>(true)
-                    .expect("Failed to init DefaultAccountState");
-                extension.state = spl_token_2022::state::AccountState::Initialized.into();
-            }
-            ExtensionType::InterestBearingConfig => {
-                let extension = mint
-                    .init_extension::<InterestBearingConfig>(true)
-                    .expect("Failed to init InterestBearingConfig");
-                extension.rate_authority = Default::default();
-                extension.initialization_timestamp = 0i64.into();
-                extension.pre_update_average_rate = 0i16.into();
-                extension.last_update_timestamp = 0i64.into();
-                extension.current_rate = 0i16.into();
-            }
-            ExtensionType::MetadataPointer => {
-                let extension = mint
-                    .init_extension::<MetadataPointer>(true)
-                    .expect("Failed to init MetadataPointer");
-                extension.authority = Default::default();
-                extension.metadata_address = Default::default();
-            }
-            ExtensionType::GroupPointer => {
-                let extension = mint
-                    .init_extension::<GroupPointer>(true)
-                    .expect("Failed to init GroupPointer");
-                extension.authority = Default::default();
-                extension.group_address = Default::default();
-            }
-            ExtensionType::GroupMemberPointer => {
-                if let Ok(extension) = mint.init_extension::<spl_token_2022::extension::group_member_pointer::GroupMemberPointer>(true) {
-                    extension.authority = Some(solana_pubkey::Pubkey::new_unique()).try_into().unwrap();
-                    extension.member_address = Some(solana_pubkey::Pubkey::new_unique()).try_into().unwrap();
-                } else {
-                    panic!("Failed to init GroupMemberPointer");
-                }
-            }
-            ExtensionType::MintCloseAuthority => {
-                let extension = mint
-                    .init_extension::<MintCloseAuthority>(true)
-                    .expect("Failed to init MintCloseAuthority");
-                extension.close_authority = Default::default();
-            }
-            ExtensionType::PermanentDelegate => {
-                let extension = mint
-                    .init_extension::<PermanentDelegate>(true)
-                    .expect("Failed to init PermanentDelegate");
-                extension.delegate = Default::default();
-            }
-            ExtensionType::ScaledUiAmount => {
-                let extension = mint
-                    .init_extension::<spl_token_2022::extension::scaled_ui_amount::ScaledUiAmountConfig>(true)
-                    .expect("Failed to init ScaledUiAmount");
-                *extension = Default::default();
-                extension.multiplier = spl_token_2022::extension::scaled_ui_amount::PodF64::from(1.0);
-                extension.new_multiplier = spl_token_2022::extension::scaled_ui_amount::PodF64::from(1.0);
-            }
-            ExtensionType::TokenMetadata => {
-                mint.init_variable_len_extension(&TokenMetadata {
-                    update_authority: Default::default(),
-                    mint: solana_pubkey::Pubkey::new_unique(),
-                    name: String::from("Test"),
-                    symbol: String::from("TEST"),
-                    uri: String::from("https://example.com/token.json"),
-                    additional_metadata: vec![],
-                }, false).expect("Failed to init TokenMetadata");
-            }
-            ExtensionType::ConfidentialTransferMint => {
-                let extension = mint
-                    .init_extension::<spl_token_2022::extension::confidential_transfer::ConfidentialTransferMint>(true)
-                    .expect("Failed to init ConfidentialTransferMint");
-                extension.authority = Default::default();
-                extension.auto_approve_new_accounts = false.into();
-                extension.auditor_elgamal_pubkey = Default::default();
-            }
-            ExtensionType::ConfidentialTransferFeeConfig => {
-                let extension = mint
-                    .init_extension::<spl_token_2022::extension::confidential_transfer_fee::ConfidentialTransferFeeConfig>(true)
-                    .expect("Failed to init ConfidentialTransferFeeConfig");
-                extension.authority = Default::default();
-                extension.withdraw_withheld_authority_elgamal_pubkey = Default::default();
-                extension.harvest_to_mint_enabled = false.into();
-                extension.withheld_amount = Default::default();
-            }
-            ExtensionType::TokenGroup => {
-                if let Ok(extension) = mint.init_extension::<TokenGroup>(true) {
-                    *extension = TokenGroup {
-                        update_authority: Default::default(),
-                        mint: solana_pubkey::Pubkey::new_unique(),
-                        size: 0u64.into(),
-                        max_size: 100u64.into(),
-                    };
-                } else {
-                    panic!("Failed to init TokenGroup");
-                }
-            }
-            ExtensionType::TokenGroupMember => {
-                if let Ok(extension) = mint.init_extension::<TokenGroupMember>(true) {
-                    *extension = TokenGroupMember {
-                        mint: solana_pubkey::Pubkey::new_unique(),
-                        group: solana_pubkey::Pubkey::new_unique(),
-                        member_number: 0u64.into(),
-                    };
-                } else {
-                    panic!("Failed to init TokenGroupMember");
-                }
-            }
-            _ => {}
-        }
-    }
-
-    data
-}
-
-/// Get all extension types by category by discovering them automatically
-pub fn get_extensions_by_category(
-    category: ExtensionCategory,
-) -> Vec<spl_token_2022::extension::ExtensionType> {
-    (0..=u16::MAX)
-        .filter_map(|i| spl_token_2022::extension::ExtensionType::try_from(i).ok())
-        .filter(|ext| categorize_extension(*ext) == category)
-        .collect()
-}
-
-/// Helper function to check if extension list contains a specific extension type
-fn has_extension(
-    extension_types: &[spl_token_2022::extension::ExtensionType],
-    target: spl_token_2022::extension::ExtensionType,
-) -> bool {
-    extension_types.iter().any(|ext| *ext == target)
-}
-
-/// Check if a combination of extension types is valid according to Token-2022 rules
-pub fn is_valid_extension_combination(
-    extension_types: &[spl_token_2022::extension::ExtensionType],
-) -> bool {
-    use spl_token_2022::extension::ExtensionType::*;
-
-    let has_scaled_ui_amount = has_extension(extension_types, ScaledUiAmount);
-    let has_interest_bearing = has_extension(extension_types, InterestBearingConfig);
-    let has_token_group = has_extension(extension_types, TokenGroup);
-    let has_group_pointer = has_extension(extension_types, GroupPointer);
-    let has_token_group_member = has_extension(extension_types, TokenGroupMember);
-    let has_group_member_pointer = has_extension(extension_types, GroupMemberPointer);
-    let has_transfer_fee_config = has_extension(extension_types, TransferFeeConfig);
-    let has_confidential_transfer_mint = has_extension(extension_types, ConfidentialTransferMint);
-    let has_confidential_transfer_fee_config =
-        has_extension(extension_types, ConfidentialTransferFeeConfig);
-
-    // ScaledUiAmount cannot be combined with InterestBearingConfig
-    if has_scaled_ui_amount && has_interest_bearing {
-        return false;
-    }
-
-    // TokenGroup requires GroupPointer extension
-    if has_token_group && !has_group_pointer {
-        return false;
-    }
-
-    // TokenGroupMember requires GroupMemberPointer extension
-    if has_token_group_member && !has_group_member_pointer {
-        return false;
-    }
-
-    // ConfidentialTransferFeeConfig requires both TransferFeeConfig AND ConfidentialTransferMint
-    if has_confidential_transfer_fee_config
-        && !(has_transfer_fee_config && has_confidential_transfer_mint)
-    {
-        return false;
-    }
-
-    // If you have both TransferFeeConfig and ConfidentialTransferMint, you MUST have ConfidentialTransferFeeConfig
-    if has_transfer_fee_config
-        && has_confidential_transfer_mint
-        && !has_confidential_transfer_fee_config
-    {
-        return false;
-    }
-
-    true
-}
-
-/// Calculate expected ATA account size using token-2022 method
-/// Adds `ImmutableOwner` as it is always included in ATA accounts
-pub fn calculate_expected_ata_data_size(
-    mint_extensions: &[spl_token_2022::extension::ExtensionType],
-) -> usize {
-    let mut account_extensions =
-        spl_token_2022::extension::ExtensionType::get_required_init_account_extensions(
-            mint_extensions,
-        );
-
-    // ATA always includes ImmutableOwner, so include it in our comparison
-    if !account_extensions.contains(&spl_token_2022::extension::ExtensionType::ImmutableOwner) {
-        account_extensions.push(spl_token_2022::extension::ExtensionType::ImmutableOwner);
-    }
-
-    spl_token_2022::extension::ExtensionType::try_calculate_account_len::<
-        spl_token_2022::state::Account,
-    >(&account_extensions)
-    .expect("Failed to calculate account length")
-}
-
-#[cfg(any(test, feature = "std"))]
-pub fn test_extension_combination_helper(
-    extensions: &[spl_token_2022::extension::ExtensionType],
-    description: &str,
-) -> Result<(), String> {
-    use crate::{
-        size::calculate_account_size_from_mint_extensions,
-        tests::token_account_len::test_extension_utils::create_mint_data_with_extensions,
-    };
-
-    let mint_data = create_mint_data_with_extensions(extensions);
-    let expected_size = calculate_expected_ata_data_size(extensions);
-    let result = calculate_account_size_from_mint_extensions(&mint_data);
-
-    if result != Some(expected_size) {
-        return Err(format!(
-            "Extension combination failed: {}. Extensions: [...extensions...]",
-            description
-        ));
-    }
-
-    Ok(())
-}
