+++ p-ata/tests/migrated/process_create_associated_token_account.rs
@@ -2,7 +2,7 @@
 #![cfg(test)]
 
 use {
-    crate::tests::test_utils::{
+    super::super::utils::test_utils::{
         build_create_ata_instruction, calculate_account_rent,
         create_mollusk_base_accounts_with_token, create_test_mint, setup_mollusk_unified,
         setup_mollusk_with_programs, CreateAtaInstructionType, MolluskAtaSetup, MolluskTokenSetup,
