+++ p-ata/tests/token_account_len/mod.rs
@@ -1,0 +1,3 @@
+#![cfg(test)]
+
+mod test_account_length_limits;
